working on userveges function
