+++ client/api/users.js
@@ -1,19 +1,19 @@
 import request from 'superagent'
 
-// import loginUser from '../actions/index'
-// import showError from '../actions/index'
+import {loginUser} from '../actions/index'
+import {showError} from '../actions/index'
 
-// export function getUserByName(user){
-//   return (dispatch) => {
-//     return request.post(`/api`, user)
-//     .then(res => {
-//       dispatch(loginUser(res.body))
-//     })
-//     .catch(err => {
-//       dispatch(showError(err.message))
-//     })
-//   }
-// }
+export function getUserByName(user){
+  return (dispatch) => {
+    return request.post(`/api`, user)
+    .then(res => {
+      dispatch(loginUser(res.body))
+    })
+    .catch(err => {
+      dispatch(showError(err.message))
+    })
+  }
+}
 
 // export function addToGarden(data){
 //   return request.post('/api/info')
